# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api": {
      "get": {
        "operationId": "AppController_getHello",
        "summary": "あいさつ取得",
        "description": "あいさつを取得する。",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ApiCommonResponseOfGetHelloResponse",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "$ref": "#/components/schemas/GetHelloResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/error": {
      "get": {
        "operationId": "AppController_getError",
        "summary": "(強制エラー)",
        "description": "強制的にエラーを出力する。",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "api example",
    "description": "The API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "example",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CommonResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "実行結果",
            "example": true
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "レスポンス送信日時",
            "example": "2024-03-12T07:45:29.583Z"
          },
          "data": {
            "type": "object",
            "description": "データ"
          },
          "error": {
            "type": "string",
            "description": "エラー内容",
            "example": []
          }
        },
        "required": [
          "success",
          "date",
          "data",
          "error"
        ]
      },
      "GetHelloResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "メッセージ",
            "example": "Hello World!"
          }
        },
        "required": [
          "message"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "description": "password",
            "example": "password"
          },
          "name": {
            "type": "string",
            "description": "name",
            "example": "Test User"
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "description": "password",
            "example": "password"
          },
          "name": {
            "type": "string",
            "description": "name",
            "example": "Test User"
          }
        }
      }
    }
  }
}